+++ src/main/java/ru/yandex/practicum/filmorate/service/RatingMpaService.java
@@ -19,7 +19,7 @@ public class RatingMpaService {
         return storage.findAll();
     }
 
-    public RatingMpa getRatingMpaById(int id) {
+    public RatingMpa getRatingMpaById(long id) {
         log.debug("getRatingMpaById {}", id);
         return storage.findOneById(id)
                 .orElseThrow(() -> new NotFoundException("ratingMpa not found"));
