+++ src/main/java/ru/yandex/practicum/filmorate/storage/interfaces/FilmGenresStorage.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.storage.interfaces;
 
 public interface FilmGenresStorage {
-    int saveFilmGenre(long filmId, int genreId);
+    int saveFilmGenre(long filmId, long genreId);
 
     int deleteFilmGenres(long filmId);
 }
